+++ src/onegov.directory/onegov/directory/archive.py
@@ -9,7 +9,7 @@ from onegov.core.csv import convert_xls_to_csv
 from onegov.core.csv import CSVFile
 from onegov.core.custom import json
 from onegov.core.utils import Bunch, rchop, is_subpath
-from onegov.directory.errors import MissingColumnError
+from onegov.directory.errors import MissingColumnError, MissingFileError
 from onegov.directory.models import Directory, DirectoryEntry
 from onegov.directory.types import DirectoryConfiguration
 from onegov.file import File
@@ -41,10 +41,13 @@ class FieldParser(object):
         # potentially be used to access files on the local system
         assert '..' not in value
         assert value.count('/') == 1
-        assert value.startswith(field.id + '/')
+        assert not value.startswith('/')
 
         path = self.archive_path / value
         assert is_subpath(str(self.archive_path), str(path))
+
+        if not path.exists():
+            raise MissingFileError(value)
 
         return Bunch(
             data=object(),