+++ src/onegov.directory/onegov/directory/errors.py
@@ -19,3 +19,9 @@ class DuplicateEntryError(OnegovDirectoryError):
     def __init__(self, name, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.name = name
+
+
+class MissingFileError(OnegovDirectoryError):
+    def __init__(self, name, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.name = name